Created a function within Nestor to return the current status of isJumping in order to ensure that within NestorRunner when we call the jump function that sound is only playing when nestor is currently jumping (sound was playing whenever the space bar was pressed before, this fixes that).
Implemented a death sound, a jumping sound, and background music. Background music is within the controller. It starts whenever the game starts, and stops whenever there is a collision. It is set to unlimited cycle count so that it will automatically repeat the track when it ends. The background music volume is adjusted to be quiet so we can hear the death and jumping sounds better. Jumping sound and death sound are played within the NestorRunner class.

In order to implement sound, I had to add a library and update module-info.java to include the javafx media library. All tracks are located in resources.edu.wsu.