Created settings.fxml. Contains two sliders for music volume and sfx volume and a main menu button.
Edited menu.fxml to contain a settings button that switches to the settings.fxml view.

Created SettingsController.java which sets the MenuController sound variables to the volume that was returned from the slider via two handle events.

Edited Sound.java to retrieve sound volume variables from the GameView and set the volume of each sound to their chosen volume.

Edited MenuController to send sound variable values to the GameView when created.

Edited GameView.java to retrieve two sound variable values to send to the Sound() when created.